a2a_demo/requirements.txt
google-adk[a2a]>=1.30.0
langgraph
langchain-core
python-dotenv
uvicorn